add six.py 1.9.0
Six is a Python 2 and 3 compatibility library. It provides utility
functions for smoothing over the differences between the Python versions
with the goal of writing Python code that is compatible on both Python
versions. See the documentation for more information on what is
provided.

Six supports every Python version since 2.5. It is contained in only one
Python file, so it can be easily copied into your project.

Change-Id: Id75aa15b62c1e8bacecf79f47425e65d355e1c82
Signed-off-by: mamh <mamh@marvell.com>